Remove the assumption of an extra temporal dimension.
Previously the text had followed an assumption of an extra
dimension for time, as other texts do, but it turns out to
be unnecessary, since a scalar + vector combination transforms
similarly.

Also removed the extra generalisations at the end (was section 4)
and plan to instead introduce them throughout the first 3 sections.